installer: handle missing system/user-backup.txt and backup
If for whatever reason the old system's system-backup.txt or
user-backup.txt are missing, don't fail the installation.

(cherry picked from commit 8ea7729d17bb0e8a4f1d86b1597a4196c3421734)

+++ scripts/installer/lib/backup.sh
@@ -134,12 +134,21 @@ create_backup()
 		[ -f "$file" ] || break
 		parse_file $file "+" $1 $2
 	done
-	parse_file "$1/$SYSTEM_BACKUP_FILE" "+" $1 $2
-	parse_file "$1/$USER_BACKUP_FILE" "+" $1 $2
+	if [ -f "$1/$SYSTEM_BACKUP_FILE" ]; then
+		parse_file "$1/$SYSTEM_BACKUP_FILE" "+" $1 $2
+	fi
+
+	if [ -f "$1/$USER_BACKUP_FILE" ]; then
+		parse_file "$1/$USER_BACKUP_FILE" "+" $1 $2
+	fi
 
 	# step 2 - remove files to drop
-	parse_file "$1/$SYSTEM_BACKUP_FILE" "-" $1 $2
-	parse_file "$1/$USER_BACKUP_FILE" "-" $1 $2
+	if [ -f "$1/$SYSTEM_BACKUP_FILE" ]; then
+		parse_file "$1/$SYSTEM_BACKUP_FILE" "-" $1 $2
+	fi
+	if [ -f "$1/$SYSTEM_BACKUP_FILE" ]; then
+		parse_file "$1/$USER_BACKUP_FILE" "-" $1 $2
+	fi
 
 	apply_fixups $1 $2
 
@@ -182,6 +191,7 @@ restore_backup()
 
     # now copy the files
     for file in $1/*; do
+	    [ -e "$file" ] || continue
 	    cp -a $file $2
     done
 }
